+++ shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -3,6 +3,7 @@ package com.project.shopapp.dtos;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.Min;
 import lombok.*;
+import java.math.BigDecimal;
 
 @Data//toString
 @Getter
@@ -20,7 +21,7 @@ public class OrderDetailDTO {
     private Long productId;
 
     @Min(value = 0, message = "Price must be >= 0")
-    private Long price;
+    private BigDecimal price;
 
     @JsonProperty("number_of_product")
     @Min(value = 1, message = "Number Of Product must be >= 1")
@@ -28,7 +29,7 @@ public class OrderDetailDTO {
 
     @Min(value = 0, message = "Total Money must be >= 0")
     @JsonProperty("total_money")
-    private Float totalMoney;
+    private BigDecimal totalMoney;
 
     private String color;
 }
